backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
xgboost==2.0.3
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
yfinance==0.2.28
ta==0.10.2
python-multipart==0.0.6
cors-middleware==0.1.0
requests==2.31.0
joblib==1.3.2
